update homeComponent to check confirm before delete user, registerComponent to handle error.message display

+++ Front-end/src/app/home/home.component.ts
@@ -1,14 +1,12 @@
 import { Component, OnInit } from '@angular/core';
 import { TokenStorageService } from '../auth/token-storage.service';
 import { UserService } from '../services/user.service';
-// import { SignUpInfo } from '../auth/signup-info';
 @Component({
   selector: 'app-home',
   templateUrl: './home.component.html',
   styleUrls: ['./home.component.scss']
 })
 export class HomeComponent implements OnInit {
-  // user = new SignUpInfo();
   message: string;
   constructor(private token: TokenStorageService,
               private usersService: UserService) { }
@@ -28,14 +26,15 @@ export class HomeComponent implements OnInit {
   }
 
   delete(): void {
-    alert("vous êtes sur le point de supprimer votre compte");
-    this.usersService.deleteUser(parseInt(this.token.getUserId())).subscribe(
-      (response) => {
-        console.log(response);
-        alert("Unsubscribe success, seen you next time");
-        this.token.signOut();
-        window.location.reload();
-      }
-    )
+    if(confirm("Do you realy want to continue this process?")){
+      this.usersService.deleteUser(parseInt(this.token.getUserId())).subscribe(
+        (response) => {
+          console.log(response);
+          alert("Unsubscribe success, seen you next time");
+          this.token.signOut();
+          window.location.reload();
+        }
+      )
+    }
   }
 }
